Improve PublicException exception message format
The previous format used () for `publicDetail` and [] for
`internalDetail` in the exception message. This made it more difficult
to distinguish between the two. Also, some developers use [] to mark
variables in string interpolation, which makes the use of [] for
`internalDetail` confusing.

So we now instead use a " - " separator for `publicDetail`, which makes
it clearer that it is public alongside `publicMessage`. And we now use
() for `internalDetail`, to avoid the confusion that [] may bring.